chore(inputs.docker): Cleanup code
